Introduce CPath and CPathExp object
A CPath is an absolute path to a specific instances credentials inside
the credential hierarchy while a CPathExp represents a regular
expression path for matching many different credential paths.

These objects can be used for validating, parsing, and working with
credential path ways in a uniform manner across the CLI and Registry.

Related arigatomachine/cli#268