All object type notation checking is toLowerCase()
Currently, there is a bug where if an id type value contained an
uppercase letter it would never validate using `utils#validate` or
`utils#type` as there wouldn't be a direct match.

This pull request fixes this bug and adds several tests to prevent this
developer error from occurring in the future.

Related arigatomachine/cli#246