Encrypt the master password with the user's passphrase
To allow for future offline access to the master key, encrypt it
directly with the user's passphrase.

Previously, we were using a key derived from the passphrase and a salt.
The salt was stored on the server, meaning we'd need to connect to the
server to get the salt before being able to derive the key.